+++ src/gas/PriceAnomalyChecker.ts
@@ -15,7 +15,7 @@ export class PriceAnomalyChecker {
     const key = `${tm.getCoinName()}-pump-dump-start`
     const anomalyStartPrice = CacheProxy.get(key)
 
-    if (tm.priceGoesUpStrong() || tm.priceGoesDownStrong()) {
+    if (tm.priceGoesStrongUp() || tm.priceGoesStrongDown()) {
       Log.debug(`${tm.getCoinName()} price anomaly detected`)
       CacheProxy.put(key, anomalyStartPrice || tm.prices[0].toString(), 120) // 2 minutes
       return PriceAnomaly.TRACKING
@@ -33,12 +33,20 @@ export class PriceAnomalyChecker {
     }
 
     if (+anomalyStartPrice > tm.currentPrice) {
-      Log.alert(`${tm.getCoinName()} price dumped for ${percent}%: ${anomalyStartPrice} -> ${tm.currentPrice}`)
+      Log.alert(
+        `📉 ${tm.getCoinName()} price dumped for ${percent}%: ${anomalyStartPrice} -> ${
+          tm.currentPrice
+        }`,
+      )
       return PriceAnomaly.DUMP
     }
 
     if (+anomalyStartPrice < tm.currentPrice) {
-      Log.alert(`${tm.getCoinName()} price pumped for ${percent}%: ${anomalyStartPrice} -> ${tm.currentPrice}`)
+      Log.alert(
+        `📈 ${tm.getCoinName()} price pumped for ${percent}%: ${anomalyStartPrice} -> ${
+          tm.currentPrice
+        }`,
+      )
       return PriceAnomaly.PUMP
     }