+++ src/gas/Process.ts
@@ -2,14 +2,17 @@ import { V2Trader } from "./Trader"
 import { Exchange } from "./Exchange"
 import { Statistics } from "./Statistics"
 import { DeadlineError, DefaultStore } from "./Store"
-import { Scores } from "./Scores"
 import { Log } from "./Common"
+import { ScoreTrader } from "./ScoreTrader"
+import { CacheProxy } from "./CacheProxy"
+import { IScores } from "./Scores"
 
 export class Process {
   static tick() {
     const store = DefaultStore
     const exchange = new Exchange(store.getConfig())
     const trader = new V2Trader(store, exchange, new Statistics(store))
+    const scores = global.TradingHelperScores.create(CacheProxy, DefaultStore, exchange) as IScores
 
     store.getTradesList().forEach((trade) => {
       try {
@@ -27,15 +30,24 @@ export class Process {
     try {
       trader.updateStableCoinsBalance()
     } catch (e) {
-      Log.error(new Error(`Failed to read stable coins balance: ${e.message}`))
+      Log.alert(`Failed to read stable coins balance`)
+      Log.error(e)
+    }
+
+    try {
+      scores.update()
+    } catch (e) {
+      Log.alert(`Failed to update scores`)
+      Log.error(e)
+    }
+
+    try {
+      new ScoreTrader(store, scores).trade()
+    } catch (e) {
+      Log.alert(`Failed to trade recommended coins`)
+      Log.error(e)
     }
 
     store.dumpChanges()
-
-    try {
-      new Scores(store, exchange).updateScores()
-    } catch (e) {
-      Log.info(e)
-    }
   }
 }