+++ src/shared-lib/responses.ts
@@ -5,3 +5,4 @@ export type AssetsResponse = {
   stableCoins: Coin[]
   trades: TradeMemo[]
 }
+